requirements.txt
numpy==1.26.4
packaging==24.1
torch==2.2.2
beautifulsoup4==4.12.3
bitsandbytes==0.43.1
chromadb==0.5.13
deepspeed==0.14.1
einops==0.8.0
PyMuPDF==1.24.4
FlagEmbedding==1.2.10
pandas==2.2.2
playwright==1.44.0
pytorch_lightning==1.9.5
PyYAML==6.0.1
rwkv==0.8.26
streamlit==1.35.0
transformers==4.42.3
triton==2.2.0
tools==0.1.9
pytils==0.4.1
pillow==10.3.0
pytesseract